requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai==1.50.0
python-docx==1.1.0
python-pptx==1.0.2
openpyxl==3.1.5
matplotlib==3.9.0
numpy==1.26.4
pydantic==2.9.0
pydantic-settings==2.5.0
